improve changelog
That was a fix for git remotes that use an url with a user name that is URL encoded

For example, "ssh://%6Aoey@localhost/path"

Of course, this is more likely to occur if the username is not all ASCII
like mine is.

And it could also happen when a scp-style url's username contains a non-ASCII
character, which gets URL encoded when git-annex internally converts that
to a ssh:// url.