revert man page changes
Revert "Linked to discussion on caveat"

This reverts commit 9fe60062a38228594ce8d48bbe1b14532934f22d.

We don't link from man pages to forum discussions. If there is a
problem, it should be fixed, and if there is a wart it should be
documented on the man page in enough detail to understand on its own.

In this case, I don't know that there is any problem at all.